DocumentForensicsAI/requirements.txt

flask
pillow
werkzeug
